Many Changes

+++ sys/kern/queue.c
@@ -2,7 +2,7 @@
 #include <kern/list.h>
 #include <kern/queue.h>
 
-void queue_init(struct queue_head *hdr, int len) {
+void queue_init(struct queue_head *hdr, size_t len) {
   hdr->free = len;
   list_init(&hdr->list);
 }
